requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
streamlit==1.28.1
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
scikit-learn==1.3.2
seaborn==0.13.0
matplotlib==3.8.2

